+++ frontend/src/__tests__/Footer.test.tsx
